Created a new HardwareClasses file that follows the example "Building a CENTERSTAGE Autonomous" on the RoadRunner guide website: https://rr.brott.dev/docs/v1-0/guides/centerstage-auto/ I only brought in the parts about creating classes for each hardware component that an autonomous roadrunner opMode would then implement. This is similar to the RobotHardware concept file we used before but using more of a class/object structure to organize the code. It will need to be adapted to take specific target positions, have incrementing functions and presets, etc. I would also like to consider a state machine structure like the RobotHardware code included.
The idea is that the Roadrunner Autonomous opMode and the teleOpModes will be both able to refer to this HardwareClasses file so that when hardware changes are made in the future it will need to only be changed in one file. I'm not sure yet how tthe actions will be used in teleOp yet though.